backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
yfinance==0.2.48
alpha_vantage==2.3.1
requests==2.32.3
ta==0.11.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
redis==5.0.1
joblib==1.4.2
prophet==1.1.6
pandas-datareader==0.10.0
# Optional deep learning - install only if needed
# tensorflow==2.16.1